Add support for motorcycle parkings (#6)
Allow a parking spot to be archived and unarchived. Archived parking spots are not eligible for new reservations, but they are still retained by the system for reporting purposes.

Add a an allowed vehicle type for each parking, defaulting to cars. When placing a reservation, the user is allowed to select only parking spots that are compatible with his vehicle.

Combining these two features allows an administrator to gracefully take a parking spot out of the system, and convert it into other parking spots for other kinds of vehicles.